Animation script v1
A script that works (!), for animating animat + falling blocks for a specified trial, agent, number.

Still, I wonder whether the block/animat position is correct, but additional experiments (using the 'perfect runs') could possibly verify this.

Also, this definitely needs tweaking + cleaning. And the files should be saved with a timestamp + the parameters (t, a, r).